requirements.txt
diffusers==0.30.3
numpy==1.24.3
opencv-python==3.4.18.65
pycryptodome==3.20.0
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
scikit-learn==1.4.2
scipy==1.11.3